Code seems to be ready; not debugged; I do not have the SSD1306
Code is ready to test all four milestones of HW 7. However, I did not pick up the chip because I was not in class at the time, so I cannot test the code until i pick it up from Prof. Marchuk (spares were not available in the lab off hours).