Uploaded HW12 Assignment Files
I do not have the physical camera component for wiring yet; all cameras are in use, and extras are unavailable.

Additionally, I could not find an example Makefile to include with the code, so I am unsure of how to proceed until I find both a camera to wire into my board and a Makefile to compile the C code properly.