Output Reviewed in Terminal, Accepted (?)
Accelerations, X, Y, and Z, look appropriate. Note: The values listed as X and Z, according to the chip and its registers, were swapped by me in the output. I did this because the chip is oriented vertically instead of flat, so swapping the values for the output makes more intuitive sense.

Gyroscope values all look off. They all read ~500 when no rotational movement occurs. When rotational movement occurs, the corresponding values approach 0. Unsure of where the error is; I didn't set any values that might change the output of the gyroscope, so I'm unsure how to proceed. After a little bit of search, I may proceed with these numbers if I cannot solve the problem since it does not truly affect the output for *this* assignment (will hopefully fix the problem after submitting this assignment, so the chip can work properly for future assignments).

Temperature value, while not correct, works as intended; when I press my finger to it or blow hot breath on it, the value increments upwards.