Imported display code, created drawLine() func
Using the necessary files and code from HW 7, all code needed to use the OLED display *should* be imported and ready for use. Everything compiled correctly, so ideally the code will work too (yet to be tested).

Additionally, I wrote up the drawLine() function to be able to create lines with a starting point, length (in pixels), and angle (in degrees).